+++ app/Http/Controllers/Controller.php
@@ -9,8 +9,5 @@ use Illuminate\Routing\Controller as BaseController;
 
 class Controller extends BaseController
 {
-    public function login(){
-
-        return view('auth.login');
-    }
+    //
 }
