+++ app/Http/Controllers/LoginController.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace App\Http\Controllers;
+
+use App\Http\Controllers\Controller;
+use Illuminate\Http\Request;
+
+class LoginController extends Controller
+{
+    public function login(){
+
+        return view('auth.login');
+        // return route('user.login');
+    }
+}
