+++ resources/views/auth/login.blade.php
@@ -4,26 +4,8 @@
 <html lang="en">
 <!--<![endif]-->
 <head>
-	<meta charset="utf-8" />
-	<title>Voucher Management Platform | Login</title>
-	<meta content="width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=no" name="viewport" />
-	<meta content="" name="description" />
-	<meta content="" name="author" />
-	
-	<!-- ================== BEGIN BASE CSS STYLE ================== -->
-	<link href="http://fonts.googleapis.com/css?family=Open+Sans:300,400,600,700" rel="stylesheet" />
-	<link href="assets/plugins/jquery-ui/jquery-ui.min.css" rel="stylesheet" />
-	<link href="assets/plugins/bootstrap/4.0.0/css/bootstrap.min.css" rel="stylesheet" />
-	<link href="assets/plugins/font-awesome/5.0/css/fontawesome-all.min.css" rel="stylesheet" />
-	<link href="assets/plugins/animate/animate.min.css" rel="stylesheet" />
-	<link href="assets/css/default/style.min.css" rel="stylesheet" />
-	<link href="assets/css/default/style-responsive.min.css" rel="stylesheet" />
-	<link href="assets/css/default/theme/default.css" rel="stylesheet" id="theme" />
-	<!-- ================== END BASE CSS STYLE ================== -->
-	
-	<!-- ================== BEGIN BASE JS ================== -->
-	<script src="assets/plugins/pace/pace.min.js"></script>
-	<!-- ================== END BASE JS ================== -->
+    @extends('components.meta')
+    @section('title', "Voucher Management Platform | Login")
 </head>
 <body class="pace-top bg-white">
 	<!-- begin #page-loader -->
@@ -93,26 +75,7 @@
                 
 	</div>
 	<!-- end page container -->
-	
-	<!-- ================== BEGIN BASE JS ================== -->
-	<script src="assets/plugins/jquery/jquery-3.2.1.min.js"></script>
-	<script src="assets/plugins/jquery-ui/jquery-ui.min.js"></script>
-	<script src="assets/plugins/bootstrap/4.0.0/js/bootstrap.bundle.min.js"></script>
-	<!--[if lt IE 9]>
-		<script src="assets/crossbrowserjs/html5shiv.js"></script>
-		<script src="assets/crossbrowserjs/respond.min.js"></script>
-		<script src="assets/crossbrowserjs/excanvas.min.js"></script>
-	<![endif]-->
-	<script src="assets/plugins/slimscroll/jquery.slimscroll.min.js"></script>
-	<script src="assets/plugins/js-cookie/js.cookie.js"></script>
-	<script src="assets/js/theme/default.min.js"></script>
-	<script src="assets/js/apps.min.js"></script>
-	<!-- ================== END BASE JS ================== -->
-
-	<script>
-		$(document).ready(function() {
-			App.init();
-		});
-	</script>
+    
+    @extends('components.external_script')
 </body>
 </html>
