+++ resources/views/components/external_script.blade.php
@@ -1,0 +1,20 @@
+<!-- ================== BEGIN BASE JS ================== -->
+<script src="assets/plugins/jquery/jquery-3.2.1.min.js"></script>
+<script src="assets/plugins/jquery-ui/jquery-ui.min.js"></script>
+<script src="assets/plugins/bootstrap/4.0.0/js/bootstrap.bundle.min.js"></script>
+<!--[if lt IE 9]>
+	<script src="../assets/crossbrowserjs/html5shiv.js"></script>
+	<script src="../assets/crossbrowserjs/respond.min.js"></script>
+	<script src="../assets/crossbrowserjs/excanvas.min.js"></script>
+<![endif]-->
+<script src="assets/plugins/slimscroll/jquery.slimscroll.min.js"></script>
+<script src="assets/plugins/js-cookie/js.cookie.js"></script>
+<script src="assets/js/theme/default.min.js"></script>
+<script src="assets/js/apps.min.js"></script>
+<!-- ================== END BASE JS ================== -->
+
+<script>
+	$(document).ready(function() {
+		App.init();
+	});
+</script>
