+++ resources/views/components/meta.blade.php
@@ -1,0 +1,21 @@
+<meta charset="UTF-8">
+<meta name="viewport" content="width=device-width, initial-scale=1.0">
+<meta http-equiv="X-UA-Compatible" content="ie=edge">
+<meta content="" name="description" />
+<meta content="" name="author" />
+<title>@yield("title")</title>
+
+<!-- ================== BEGIN BASE CSS STYLE ================== -->
+<link href="http://fonts.googleapis.com/css?family=Open+Sans:300,400,600,700" rel="stylesheet" />
+<link href="assets/plugins/jquery-ui/jquery-ui.min.css" rel="stylesheet" />
+<link href="assets/plugins/bootstrap/4.0.0/css/bootstrap.min.css" rel="stylesheet" />
+<link href="assets/plugins/font-awesome/5.0/css/fontawesome-all.min.css" rel="stylesheet" />
+<link href="assets/plugins/animate/animate.min.css" rel="stylesheet" />
+<link href="assets/css/default/style.min.css" rel="stylesheet" />
+<link href="assets/css/default/style-responsive.min.css" rel="stylesheet" />
+<link href="assets/css/default/theme/default.css" rel="stylesheet" id="theme" />
+<!-- ================== END BASE CSS STYLE ================== -->
+
+<!-- ================== BEGIN BASE JS ================== -->
+<script src="assets/plugins/pace/pace.min.js"></script>
+<!-- ================== END BASE JS ================== -->
