+++ resources/views/pages/supplier_registration/index.blade.php
